requirements.txt
# Core dependencies
click>=8.1.0
gitpython>=3.1.40
PyGithub>=2.1.1
atlassian-python-api>=3.41.0
detect-secrets>=1.4.0
pyyaml>=6.0
cryptography>=41.0.0
pydantic>=2.5.0
python-dotenv>=1.0.0
colorama>=0.4.6
rich>=13.7.0
requests>=2.31.0
aiohttp>=3.9.0
PyNaCl>=1.5.0  # For GitHub secrets encryption

# API dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
alembic>=1.13.0
psycopg2-binary>=2.9.9
redis>=5.0.1
celery>=5.3.4
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
